+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Add extension to create team site with no group using SPSiteManager #2364 [NicolajHedeager]
 
 ### Changed