+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## [3.12.1908.0 - August 2019 release]
+## [unreleased]
 
 ### Added
 
@@ -14,9 +14,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Fix: Handle SPSiteManager/Create with SiteStatus = 1 (provisioning status)
+- Fix: Avoid 'object reference not found' error [kachihro]
+
+## [3.12.1908.0 - August 2019 release]
+
+### Added
+
+### Changed
+
 - Fix: Added Built in CT Html Page layout #2321 [SchauDK]
 - Fix: Handle GroupSiteManager/CreateGroupEx with SiteStatus = 1 (provisioning status)
-- Fix: Handle SPSiteManager/Create with SiteStatus = 1 (provisioning status)
 - Fix: Increase default timeout for ALM API calls
 
 ## [3.11.1907.0 - July 2019 release]